refactor(config): replace figment with configured
