+++ Cargo.toml
@@ -14,12 +14,9 @@ publish       = false
 anyhow      = { version = "1.0", features = [ "backtrace" ] }
 api-version = { version = "0.3" }
 axum        = { version = "0.8", features = [ "http2", "json" ] }
-figment     = { version = "0.10", features = [ "env", "yaml" ] }
+configured  = { version = "0.9" }
 log         = { version = "0.4" }
 logforth    = { version = "0.30", features = [ "starter-log" ] }
 serde       = { version = "1.0", features = [ "derive" ] }
 tokio       = { version = "1", features = [ "macros", "rt-multi-thread", "signal" ] }
 tower       = { version = "0.5" }
-
-[dev-dependencies]
-assert_matches = { version = "1.5" }
