+++ src/main.rs
@@ -1,18 +1,15 @@
-mod config;
 mod infra;
-mod telemetry;
 
 use anyhow::{Context, Result};
+use configured::{Case, Configured};
 use log::{error, info};
+use logforth::{append::Stdout, filter::rustlog::RustLogFilterBuilder, layout::JsonLayout};
 use serde::Deserialize;
 use std::panic;
 
-use crate::config::ConfigExt;
-
 #[tokio::main]
 async fn main() {
-    // Initialize logging.
-    telemetry::init_logging();
+    init_logging();
 
     // Replace the default panic hook with one that uses structured logging at ERROR level.
     panic::set_hook(Box::new(|panic| error!(panic:%; "process panicked")));
@@ -31,8 +28,18 @@ struct Config {
 }
 
 async fn run() -> Result<()> {
-    let config = Config::load().context("load configuration")?;
+    let config = Config::load(Case::Snake).context("load configuration")?;
     info!(config:?; "starting");
 
     infra::api::serve(config.infra.api).await
+}
+
+pub fn init_logging() {
+    logforth::starter_log::builder()
+        .dispatch(|dispatch| {
+            dispatch
+                .filter(RustLogFilterBuilder::from_default_env().build())
+                .append(Stdout::default().with_layout(JsonLayout::default()))
+        })
+        .apply();
 }
